Pull logic out of React
