+++ package.json
@@ -15,14 +15,10 @@
     "node": ">=10"
   },
   "scripts": {
-    "start": "tsdx watch",
     "build": "tsdx build",
-    "test": "tsdx test --passWithNoTests",
-    "prepare": "tsdx build",
+    "test": "vitest",
     "size": "size-limit",
-    "analyze": "size-limit --why",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "analyze": "size-limit --why"
   },
   "peerDependencies": {
     "react": ">=16"
@@ -51,17 +47,19 @@
   "devDependencies": {
     "@size-limit/preset-small-lib": "^7.0.8",
     "@testing-library/react-hooks": "^8.0.1",
-    "@types/react": "^17.0.43",
-    "@types/react-dom": "^17.0.14",
+    "@testing-library/vue": "^7.0.0",
+    "@types/react": "*",
+    "happy-dom": "^10.0.3",
     "husky": "^7.0.4",
     "jotai": ">=1",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
-    "react-is": "^18.0.0",
-    "react-test-renderer": "^18.0.0",
+    "react": "*",
+    "react-is": "*",
+    "react-test-renderer": "*",
     "size-limit": "^7.0.8",
-    "tsdx": "^0.14.1",
     "tslib": "^2.3.1",
-    "typescript": "^4.6.3"
+    "tsup": "^7.1.0",
+    "typescript": "^4.6.3",
+    "vitest": "^0.33.0",
+    "vue": "^3.3.4"
   }
 }
