+++ src/vanilla/scope.ts
@@ -1,3 +1,5 @@
+import { ScopeTuple } from "./types";
+
 export type MoleculeScope<T> = {
   defaultValue: T;
   displayName?: string;
@@ -10,4 +12,24 @@ export function createScope(defaultValue?: unknown): MoleculeScope<unknown> {
   return {
     defaultValue,
   };
+}
+
+
+export function getDownstreamScopes(parentScopes: ScopeTuple<unknown>[], nextTuple: ScopeTuple<unknown>) {
+  const [scope] = nextTuple;
+  const found = parentScopes.findIndex((scopeTuple) => {
+    const foundScope = scopeTuple[0];
+    return foundScope === scope;
+  });
+
+  const downstreamScopes = found >= 0
+    ? // Replace inline (when found)
+    [
+      ...parentScopes.slice(0, found),
+      nextTuple,
+      ...parentScopes.slice(found + 1),
+    ]
+    : // Append to the end (when not found)
+    [...parentScopes, nextTuple];
+  return downstreamScopes;
 }
